move Basic auth to new Hook
